requirements.txt
aniso8601==10.0.1
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
passlib==1.7.4
pytz==2025.2
setuptools==80.3.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
uuid==1.30
Werkzeug==3.1.3
WTForms==3.2.1
celery[Redis]
